+++ lib/core/step.ts
@@ -1,7 +1,8 @@
 import TerminalLog from "../utils/log.utils";
 import { randomUUID } from "node:crypto";
-import { Locale, LocalOpt, Method, SProps, Steps, Type } from "../types";
-import { Messages } from "../utils";
+import { CatchError, CatchProps, Locale, LocalOpt } from "../types";
+import { Method, SProps, Steps, Type } from "../types";
+import { Messages, stepPropsFromAxiosError } from "../utils";
 
 export class Step implements Steps {
     readonly uid!: string;
@@ -94,6 +95,17 @@ export class Step implements Steps {
      */
     public static create(props: Partial<SProps>): Readonly<Step> {
         return new Step({ ...props })
+    }
+
+    /**
+     * @description Create an step instance from error: Error from a try catch block.
+     * @param error as instance of Error from catch block.
+     * @param props object with CatchProps.
+     */
+    public static catch(error: Error | CatchError, props?: CatchProps): Readonly<Step> {
+        const rmKeys = (props && props.remove) ? props.remove : [];
+        const params = stepPropsFromAxiosError(error as CatchError, rmKeys);
+        return Step.create(params);
     }
 
     /**