+++ lib/index.ts
@@ -8,4 +8,5 @@ export * from './utils/ensure-local-folder.util';
 export * from './utils/get-folder-name.util';
 export * from './utils/get-file-name.util';
 export * from './utils/get-file-name.util';
+export * from './utils/step-from-axios-error.util';
 export * from './types';